oscar-backend/requirements.txt
alembic==1.18.4
asyncpg==0.31.0
fastapi==0.135.1
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
psycopg2-binary==2.9.11
pydantic-settings==2.13.1
python-dotenv==1.2.2
email-validator==2.3.0
sendgrid==6.12.5
uvicorn==0.41.0
twilio==9.10.3
openai==2.26.0
redis==7.3.0